+++ resources/views/default/column/filter/select.blade.php
@@ -1,3 +1,7 @@
-<div {!! $width !!}>
-  {!! Form::select('', $options, $default, $attributesArray) !!}
+<div {!! $width !!} class="adm-form-select">
+    <x-sleepingowl::form.select
+        :name="''"
+        :options="$options"
+        :value="$default"
+        :attributes="$attributesArray" />
 </div>