+++ resources/views/default/column/url.blade.php
@@ -1,7 +1,17 @@
 <div {!! $attributes !!}>
 	@if ($visibled)
+        @php
+            $attrs = '';
+            foreach ($linkAttributes as $attr => $val) {
+                if (is_bool($val)) {
+                    if ($val) $attrs .= ' ' . $attr;
+                } else {
+                    $attrs .= ' ' . $attr . '="' . e($val) . '"';
+                }
+            }
+        @endphp
 		@if (!empty($value))
-			<a href="{{ $value }}" {{ app('html')->attributes($linkAttributes) }} target="_blank">
+			<a href="{{ $value }}" {!! $attrs !!} target="_blank">
 				@if($icon)
 					<i class="{{ $icon }}" data-toggle="tooltip" title="{{ trans('sleeping_owl::lang.table.filter-goto') }}"></i>
 				@endif