+++ resources/views/default/form/element/select2.blade.php
@@ -7,8 +7,21 @@
         @endif
     </label>
 
-    <div>
-        {!! Form::select($name, collect($options)->pluck('text', 'id'), $value, $attributes_array) !!}
+    <div class="adm-form-select2">
+
+            @php
+//            dd($attributes, $attributes_array);
+//                /**
+//                * @nit Daan Проверить зачем тут был pluck
+//                :options="collect($options)->pluck('text', 'id')"
+//                */
+            @endphp
+
+        <x-sleepingowl::form.select
+            :name="$name"
+            :options="$options"
+            :value="$value"
+            :attributes="$attributes_array" />
     </div>
 
     @include(AdminTemplate::getViewPath('form.element.partials.helptext'))