+++ src/Display/Column/Checkbox.php
@@ -2,7 +2,7 @@
 
 namespace SleepingOwl\Admin\Display\Column;
 
-use Form;
+//use Form;
 
 class Checkbox extends NamedColumn
 {
@@ -24,9 +24,7 @@ class Checkbox extends NamedColumn
     public function __construct($label = null)
     {
         parent::__construct($label);
-        $this->setLabel(
-            Form::checkbox(null, 1, false, ['class' => 'adminCheckboxAll']
-            ));
+        $this->setLabel('<input type="checkbox" class="adminCheckboxAll" value="1">');
     }
 
     /**